refactor. make some reusable components
